requirements.txt
# 显微镜控制系统 - Python依赖包
# Generated for Raspberry Pi 5 environment
# 基于实际安装的包版本

# Web框架和实时通信
Flask==2.2.2
Flask-SocketIO==5.5.1
python-socketio==5.13.0
python-engineio==4.12.2

# 图像处理和计算机视觉
# opencv-python==4.8.1.78  # 使用系统预装版本或手动安装
Pillow==9.4.0
numpy==1.24.2

# 树莓派摄像头支持
picamera2==0.3.27

# 硬件控制 (树莓派GPIO和传感器)
RPi.GPIO==0.7.1
adafruit-circuitpython-ads1x15==2.4.4
adafruit-blinka==8.66.0
Adafruit-ADS1x15==1.0.2
Adafruit-GPIO==1.0.3
Adafruit-PureIO==1.1.11

# 数据分析和科学计算
scipy==1.10.1

# 进度条显示
tqdm==4.64.1

# Web服务器支持
Werkzeug==2.2.2

# 网络和HTTP客户端
requests==2.28.1

# 系统信息获取
netifaces==0.10.6

# 实时通信支持
bidict==0.23.1
simple-websocket==1.1.0
wsproto==1.2.0

# 其他依赖
click==8.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2

# VLM大模型支持
openai>=1.0.0
python-dotenv>=1.0.0

# 注释：Python内置模块，无需安装
# json, multiprocessing, threading, time, os, io, base64, subprocess, tempfile, re, pathlib 